+++ src/swganh/tre/visitors/terrain/detail/boundary_polyline.h
@@ -14,13 +14,13 @@ namespace tre
 	public:
 
 		virtual void Deserialize(anh::ByteBuffer& buffer);
-		virtual bool IsContained(double px, double pz);
-		virtual double Process(double px, double pz);
+		virtual bool IsContained(float px, float pz);
+		virtual float Process(float px, float pz);
 		
 	protected:
 		std::vector<glm::vec2> verts;
-		double line_width;
-		double max_x, max_z, min_x, min_z;
+		float line_width;
+		float max_x, max_z, min_x, min_z;
 	};
 
 }
