+++ src/swganh/tre/visitors/terrain/detail/boundary_rectangle.cc
@@ -14,7 +14,7 @@ void BoundaryRectangle::Deserialize(anh::ByteBuffer& buffer)
 	feather_amount = buffer.read<float>();
 
 	//Make sure x1, z1 is lower left and x2,z2 is top right
-	double temp;
+	float temp;
 	if (x1 > x2)
 	{
 		temp = x1;
@@ -30,7 +30,7 @@ void BoundaryRectangle::Deserialize(anh::ByteBuffer& buffer)
 	}
 }
 
-bool BoundaryRectangle::IsContained(double px, double pz)
+bool BoundaryRectangle::IsContained(float px, float pz)
 {
 	if (x2 >= px && x1 <= px && z2 >= pz && z1 <= pz)
 		return true;
@@ -38,31 +38,31 @@ bool BoundaryRectangle::IsContained(double px, double pz)
 	return false;
 }
 
-double BoundaryRectangle::Process(double px, double pz)
+float BoundaryRectangle::Process(float px, float pz)
 {
-	double result;
+	float result;
 	
 	if (!IsContained(px, pz))
 		result = 0.0f;
 	else
 	{
-		double min_distx = px - x1;
-		double max_distx = x2 - px;
-		double min_distz = pz - z1;
-		double max_distz = z2 - pz;
-		double x_length = x2 - x1;
-		double length = z2 - z1;
+		float min_distx = px - x1;
+		float max_distx = x2 - px;
+		float min_distz = pz - z1;
+		float max_distz = z2 - pz;
+		float x_length = x2 - x1;
+		float length = z2 - z1;
 
 		if (x_length < length)
 			length = x_length;
 
-		double feather_length = feather_amount * length * 0.5;
-		double feather_result = feather_length;
+		float feather_length = (float)( feather_amount * length * 0.5);
+		float feather_result = feather_length;
 
-		double newX0 = x1 + feather_length;
-		double newX1 = x2 - feather_length;
-		double newZ0 = z1 + feather_length;
-		double newZ1 = z2 - feather_length;
+		float newX0 = x1 + feather_length;
+		float newX1 = x2 - feather_length;
+		float newZ0 = z1 + feather_length;
+		float newZ1 = z2 - feather_length;
 
 		if (px < newX1 || px > newX0 || pz < newZ1 || pz > newZ0)
 			return 1.0f;
